Fixed a AC bug.

+++ EarthWithMagicAPI/API/Creature/CreatureAttributes.cs
@@ -30,10 +30,10 @@
         /// <param name="initiative">Used to determine the order of creature action.</param>
         /// <param name="toHit">Used to determine if we will hit a creature with our current state of equipment.</param>
         /// <param name="dodge">The chance that the character will dodge something.</param>
-        public CreatureAttributes(Gender gender = Gender.Unspecified, Alignment alignment = Alignment.LawfulGood, Race race = Race.Human, int AC = 0, int maxHealth = 0, int health = 0, int dexterity = 0, int strength = 0,
-            int constitution = 0, int charisma = 0, int wisdom = 0, int fireResistence = 0, int acidResistence = 0,
-            int poisonResistence = 0, int electricResistence = 0, int coldResistence = 0, int magicResistence = 0, int charmResistence = 0,
-            int sleepResistence = 0, bool needsOxygen = false, int initiative = 0, int toHit = 0, int dodge = 0, int intelligence = 0)
+        public CreatureAttributes(Gender gender = Gender.Unspecified, Alignment alignment = Alignment.LawfulGood, Race race = Race.Human, double AC = 0, double maxHealth = 0, double health = 0, int dexterity = 0, int strength = 0,
+            int constitution = 0, int charisma = 0, int wisdom = 0, double fireResistence = 0, double acidResistence = 0,
+            double poisonResistence = 0, double electricResistence = 0, double coldResistence = 0, double magicResistence = 0, double charmResistence = 0,
+            double sleepResistence = 0, bool needsOxygen = false, int initiative = 0, double toHit = 0, double dodge = 0, int intelligence = 0)
         {
             this.Gender = gender;
             this.BaseAC = AC;
@@ -64,8 +64,8 @@
         /// <summary>
         /// The chance that we will successfully hit with the currently equipped weapon.
         /// </summary>
-        public int BaseToHit;
-        public int ToHit;
+        public double BaseToHit;
+        public double ToHit;
 
         /// <summary>
         /// The gender of the creature.
@@ -208,8 +208,8 @@
         /// <summary>
         /// The chance that the creature will dodge.
         /// </summary>
-        public int BaseDodge;
-        public int Dodge;
+        public double BaseDodge;
+        public double Dodge;
 
         /// <summary>
         /// The intelligence of the creature.
